requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
lxml==5.3.0
pydantic==2.10.6
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
tenacity==9.0.0
openai==1.68.2
typer==0.12.5
playwright==1.51.0
streamlit==1.43.2